backend/requirements.txt
# FastAPI and Web Framework
fastapi
uvicorn[standard]

# Database and ORM
sqlmodel
sqlalchemy
psycopg2-binary
alembic

# Data Validation and Serialization
pydantic
pydantic-settings
email-validator

# Authentication
PyJWT
cryptography
passlib[bcrypt]

# Configuration
python-dotenv

# Testing
pytest
pytest-asyncio
httpx

# Utilities
python-multipart

# Rate Limiting
slowapi

# AI Service
cerebras-cloud-sdk
